% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-gat.R
\name{gat.fit}
\alias{gat.fit}
\title{Estimate GAT parameters}
\usage{
gat.fit(x, start, lower.bound, upper.bound, control = NULL)
}
\arguments{
\item{x}{Numeric vector of observations for estimating parameters.}

\item{start}{(optional) starting values of parameters as c(mean,sd,nu,d,xi) 
for the optimization algorithm}

\item{lower.bound}{(optional) lower bounds for the optimization algorithm}

\item{upper.bound}{(optional) lower bounds for the optimization algorithm}
}
\value{
An object with the optimization output, including estimated parameters,
  convergence status, objective value, and additional diagnostic information. The
  optimization is performed using \code{\link[Rsolnp]{solnp}}.
}
\description{
Functions to estimate all parameters of the GAT distribution from a vector
of iid observations.
}
\details{
It optimizes the log-likelihood based on dgat data
}
\examples{
# simulate random values from GAT distribution
x = rgat(n = 1000, mean = 2, sd = 1, nu = 2, d = 1, xi = 3)
# estimate the parameters using the observations x
gat.fit(x)$pars

}
\author{
Thiago do Rego Sousa
}
