% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEVStableGarch-package.R
\docType{package}
\name{GEVStableGarch-package}
\alias{GEVStableGarch}
\alias{GEVStableGarch-package}
\title{GEVStableGarch: ARMA-GARCH/APARCH models with stable, GEV and GAt innovations}
\description{
Simulation, estimation and forecasting of ARMA-GARCH/APARCH models whose
innovations follow a stable (S0), GEV or GAt distribution, or any user
defined innovation distribution.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thiagodoregosousa/GEVStableGarch}
  \item Report bugs at \url{https://github.com/thiagodoregosousa/GEVStableGarch/issues}
}

}
\author{
\strong{Maintainer}: Thiago do Rego Sousa \email{thiagodoregosousa@gmail.com}

Authors:
\itemize{
  \item Thiago do Rego Sousa \email{thiagodoregosousa@gmail.com}
}

Other contributors:
\itemize{
  \item Cira Otiniano \email{cira@unb.br} [contributor]
  \item Silvia Lopes \email{silviarc.lopes@gmail.com} [contributor]
  \item Diethelm Wuertz \email{wuertz@itp.phys.ethz.ch} (filter and model structure adapted from fGarch) [contributor, copyright holder]
}

}
\keyword{internal}
