\name{show-methods}


\docType{methods}


\alias{show-methods}
\alias{show,fGEVSTABLEGARCH-method}
\alias{show,fGEVSTABLEGARCHSPEC-method}


\title{GEVSTABLEGARCH Package Show Methods}


\description{

Methods to organize the output to the user when printing objects of class GEVSTABLEGARCH

}


\section{Methods}{
\describe{

    \item{\code{signature(object = "fGEVSTABLEGARCH")}}{
        Print function for objects of class \code{"fGEVSTABLEGARCH"}
    }

    \item{\code{signature(object = "fGEVSTABLEGARCHSPEC")}}{
        Print function for objects of class \code{"fGEVSTABLEGARCHSPEC"}
    }
}}


\author{

Thiago do Rego Sousa, Cira Etheowalda Guevara Otiniano 
and Silvia Regina Costa Lopes.

}


\keyword{models}